Clarify limited logo usage rights granted by orgs
Several users have reported that securing internal permission to put their company logo on Who Uses Pants is easiest when legal/PR/etc have clarity that the usage rights granted are limited. This is an attempt to communicate more clearly with people interacting with this form and those whose authorization they seek.