Fix docker invalidation bug.
Assume a "child" image that depends on a "parent" image
(via a FROM command).

Previously, the child image building Process had no mention
of the parent image's unique ID. So, even though we would re-run
the parent building process every time (because it was set
to PER_SESSION scope in #13464), Pants would not know that the
child building process depended on the parent building process
completing first (and creating the new parent image as a side effect).

Now we write a dummy file containing the ids of all upstream images
into the context in which a child image build is executed. This
causes the child building process to invalidate correctly.

Note that we still need the PER_SESSION scope because we have
to re-run all image build processes even when there were no changes at
all, in case the user did a `docker image rm`.

This highlights the complications and dangers of interacting with
a stateful external daemon like docker.

[ci skip-rust]

[ci skip-build-wheels]