Support macOS-10.15 CI on a self-hosted runner. (#16440)
This just builds wheels, we leave full testing to the MacOS11 jobs, and so this change doesn't affect the cron workflow.

- Also removes the ARM64 cron build, which should never have been set up in the first place, since we don't run tests on M1s, just wheel builds, and those aren't run in the cron workflow for other platforms.

- Also fixes the M1 build to target a single fully-specified tag, `macOS-11-ARM64`, instead of a combination of `self-hosted`, `ARM64` and `macOS11`. This is because there is a big gotcha with tagging the runner with separate tags: Any job that targets `macOS11` alone can be picked up by that runner. 

The only reason our regular `macos-11` builds, intended to run on GitHub's hosted runners, didn't run on our self-hosted runner, is that we misspelled `macOS11` without the dash when tagging the self-hosted runner. A deliberate misspelling is not a great thing to rely on for correctness... 

Unfortunately we don't control the tags of the GitHub-hosted runners, and there is no `github-hosted` tag, similar to `self-hosted`, to differentiate the two. And in any case, we should use tags to target what we care about, namely platform and OS version, and not technicalities like who hosts the runner.

So the best solution is to target a single tag that fully specifies the distinctions we care about.