Add IBM to user list.
We already had the ability to display the company name in large type
instead of the logo, if no logo was provided. However it looks pretty
bad to then repeat the company name in small type underneath it.

In fact, since all logos include the company name as legible text,
and we had already discussed remove the name text as a separate
table row, this change does just that. The name is now used as
alt text for the logo image, if any.