Mark `[python].requirement_constraints` as deprecated (but not planned for removal) (#16606)
We decided in Slack to not plan to remove this mechanism until Pants 3.0. While we strongly encourage using `[python].resolves` instead, we don't want to _force_ users to migrate. We only want to _encourage_ migration.

This PR also refactors `pex_from_targets.py` so that the impact of the old code is even better isolated. Note that we will always need to keep around a lot of the relevant code with how to subset via `--repository-pex` because we decided to support manually managed `requirements.txt`-style locks.