The environment to use becomes a context-specific parameter (#16721)
In order to differentiate subgraphs of the `@rule` graph for different platforms and environments for #7735 (to support #13682), the environment that is in use needs to become a very widespread parameter in the `@rule` graph.

To do that, this change:
1. Adjusts all production `QueryRule`s to declare that an `EnvironmentName` must be provided.
2. Adjusts most `@union`s to use  #16717 to declare that an `EnvironmentName` will be available.
3. Computes the `Platform` from the `EnvironmentName` (currently as a noop, since we don't actually have multiple instances of the `EnvironmentName`).

A caveat for point 1 is that to avoid churn, we only update _production_ `QueryRule`s in this change. For continuity in tests, `RuleRunner` will install a singleton `EnvironmentName` by default, and uses a (probably temporary) facility to filter the `EnvironmentName` back out of all `QueryRule`s. Tests that actually want to test with multiple environments should disable that facility.