Fix equality implementation in ordered_set (#16861)
The current implementation of `_AbstractOrderedSet.__eq__` uses `zip_longest`, which fills the smaller iterable with `None`. This was probably done to handle sets of different sizes, but causes  `Ordered([1]) == Ordered([1, None])`, which is incorrect. The changed implementation first compares the length of both sets and then compares all respective pairs of elements using `zip`.

[ci skip-rust]