`Environment`, `EnvironmentRequest`, and `CompleteEnvironment` now include `Vars` in the name (#16851)
This avoids ambiguity with the new `Environment` mechanism.

We can't outright remove the old names because it breaks the Toolchain plugin used by pantsbuild/pants. At the same time, I don't know how to pull off deprecating the old aliases? Using a subclass means it would be a distinct type, which the engine does not like.

[ci skip-rust]