Introduce new schema for lint plugins (#16735)
This PR is the first of a refactor of `fmt`/`lint` which aims to simplify the implementations enough to make `fix` an easy drop-in addition and support targetless formatters/fixers.

To understand why these changes are the way they are, it's important to consider the end-state is `lint` is able to simply request all the lint results for our tools, regardless of being a linter/formatter/fixer. With that in mind, the types in `lint` need to share a common initializer interface, since we don't know exactly what we're constructing. The lowest-common-denominator is a `Snapshot`.
Additionally, formatters/fixers are run serially on files that that multiple tools touch. Therefore in order to support targetless formatters/fixers, we need to know what files they care about from the inputs. Therefore, a 2-rule process of partitioning the inputs, then running the tool.

With that in mind, this PR:
- Uproots `lint` plugins, such that they have to define 2 rules. One which goes `FooRequest.PartitionRequest -> XYZPartitions` and one that goes `FooRequest.Batch -> LintResult`.
- The `Batch` instance contains partitions which altogether form the set of inputs to be run on. It also allows for arbitrary objects to be given (provided it supports the regular node requirements) as each partition which will be then forwarded to the runner implementation. This allows `lint` plugins to do interesting things
- Makes lint plugins register their rule as `*FooRequest.registration_rules()`. This install the necessary `UnionRule`s, and will be extended for `fmt` and `fix` plugins

The way we accomplish this is by using multiple unions:
- `LintRequest`, whose members are the user-defined request types. It is membership-only (no `Get`) and used to help us find members of the following unions
- `LintRequest.Batch`: A union holding the implementors of the `Batch -> LintResult` rules
- `LintTargetsRequest.PartitionRequest`: Target-based "partitioner implementations
- `LintFilesRequest.PartitionRequest`: File-based "partitioner" implementations

Future sketches are to:
- Have a rule to convert a `FmtResult` to `LintResult`
- Have the `fmt` types inherit from the `Lint` types, and continue the union shenanigans
- Track the request types by name so the result types don't need to propagate it

Some minor perks of the new strategy (once fully implemented):
- Slightly improved perf
  - Early-exit for skipped/filtered tools
  - We batch AFTER partitions, which should yield better results
  - targetless tools early-exit if they dont apply
- Cleaner code
  - No more returning a `LintResults([LintResult(...))`. Its just a `LintResult`
  - No more `skip`/`skipped`


Also fixes #16533 :sweat_smile: 

[ci skip-rust]
[ci skip-build-wheels]