Environment sensitive options (#16840)
This adds the first pass at environment-sensitive options support as described in the [design doc](https://docs.google.com/document/d/14xYllPF_FjFMtJybTi2l_bZKfZsHvzGta7CIggdm_ms/edit#)

It replaces the existing `python_*` fields on Environment Targets with ones that are synthesised at runtime from options that have been marked as `environment_sensitive`. 

The bulk of the work is over in `environments.py`, specifically `add_option_fields_for()`, which consumes a `Subsystem` type and registers newly-generated fields that correspond to the environment-sensitive options.

There is also `get_option`, which is a temporary API for getting an option from either a `Subsystem` or an `EnvironmentTarget` -- this is ahead of an anticipated refactor where subsystems will be created per-environment.