Add `platform` field for Docker environment targets  (#16867)
Per https://github.com/pantsbuild/pants/issues/16852, we need to set the `Platform` when Docker is used. It's wrong to use the localhost.

Originally, we were thinking to inspect the docker image itself via either `docker inspect` or running `uname` inside the container. 

However, an issue is multiplatform images, like `python:3.9`, which can work with multiple platforms. We need to allow the user to disambiguate which platform they are targeting. Hence, we have a `platform` field.

By default, we match the CPU architecture of the local host, which is what Docker does. A user can otherwise hardcode to a certain platform.

[ci skip-rust]
[ci skip-build-wheels]