Load environment variables from Docker image (#16876)
Closes https://github.com/pantsbuild/pants/issues/16754.

We simply run the `env` process. We could alternatively run `docker inspect`, but this is much easier to implement and this strategy will work for remote execution environments too.

--

Note that we do _not_ strip environment variables for Docker images, unlike local processes! This is because we will be guaranteed that every user has the exact same environment, so there is no need to strip environment variables; whereas for local processes, we need to strip in an attempt to reduce Works On My Machine issues.

Still, we need to extract out the `CompleteEnvironmentVars` at the start of the run so that rules can augment and override certain values, such as the `[test].extra_env_vars` option.

This key detail is what makes our `Process` to run `env` possible.

[ci skip-rust]
[ci skip-build-wheels]