docker: set platform when creating containers (#16886)
Update `bollard` and make use of new ability to set platform upon container creation.

For tests, the `busybox` image is only available for Linux and not macOS. Thus on macOS the tests fail since no macOS version of busybox image is available. Solve by pinning the platform used for tests to Linux. On Linux this is fine since same OS. On macOS, this works due to macOS Docker support for running Linux containers.