Platform for remote execution and caching comes from Python (#16898)
It's possible to have n remote execution environments, and those may each have their own platform. So it's not safe for this value to be hardcoded in Rust.

I believe this fixes an actual bug for remote caching. We were setting the platform as the localhost, but if remote caching is used in conjunction with remote execution, that is often wrong. Specifically, using RE from macOS would store in the remote cache a macOS platform, when it should have been storing Linux.