add `shutdown` method to CommandRunner (#16930)
The Docker command runner needs to cleanly remove cached containers when it shuts down. The removal API must be called in an async context, however, so the shutdown cannot be performed from drop handler since that is a non-async context.

Add a `shutdown` method to `CommandRunner` and implement it for all relevant command runners. For the Docker command runner, this means that the shutdown logic for `ContainerCache` will be invoked in an async context.

[ci skip-build-wheels]