java resources jar file is non determinstic
At the moment, java or scala projects won't benefit much from the remote cache because the resources jar file will usually win the remote cache read. As zip is non-determinstic, which means that the resources jar file would invalidate any dependents in the graph.

Using reproducible_jars would solve that, but the price would be running the strip jar for all jars.

[ci skip-rust]

[ci skip-build-wheels]