Add an optional speculation delay when using a remote cache (#16922)
When `nailgun` is combined with a remote cache, cancellation of `nailgun` processes due to cache hits results in `nailgun` servers getting killed (as expected: see #14937 for more on that topic).

But the cost of killing a `nailgun` server (which makes _future_ processes run more slowly) is not currently accounted for by our default setting of speculating for all requests. And figuring out the exact likelihood that we will hit the cache is a challenging problem.

This change adds an (optional, but enabled by default) delay before starting a `nailgun` process when a remote cache is in use.

Fixes #16921.

[ci skip-build-wheels]