Add `fallback_environment` for `remote_environment` (#16955)
Closes https://github.com/pantsbuild/pants/issues/16907.

This feature is particularly powerful when you are using a Docker image for remote execution and you set the fallback environment is a Docker environment: it means you can always use the same container.

Note that we considered instead having consuming targets set this fallback logic in the `environment: str` field. This would mean adding a `docker_image` field to the `remote_execution` environment target where the user gives us a hint what the value is, and then call sites use a new DSL like `environment="image:centos6`. This PR was accepted in https://pantsbuild.slack.com/archives/C0D7TNJHL/p1663790578352079 as a simper design, especially because we want to keep things simpler for _consumers_ of this feature, which make up the majority of the users.

[ci skip-rust]