Remove bad linebreaks from python-run-goal page. (Cherry-pick of #16976) (#16981)
Newlines intended to break up the source, for readability, have the effect of inserting a line break in the rendered output. This may be related to how readme.com processes uploads, because markdown is not supposed to behave this way AFAIK.

Until we figure this out, we can at least manually fix pages where we notice issues.