Make more option values environment-sensitive (#16984)
This migrates _most_ of the options identified in [the design document](https://docs.google.com/document/d/1vXRHWK7ZjAIp2BxWLwRYm1QOKDeXx02ONQWvXDloxkg/edit#) for environment configuration to the new `Subsystem.EnvironmentAware` mechanism.

This includes:

* [apache-thrift]
* [python-native-code]
* [shell-setup]
* [subprocess-environment]
* [twine]
* [pex]
* [docker]
* [test] goal
* [jvm]

Please let me know if more options in your subsystem should have support for per-environment option settings

**Unaddressed from the design doc:**

Currently `[golang]` is missing, because there's a few fields beyond those in the design doc that may need to be migrated, and I am awaiting confirmation. If those questions are resolved before this is ready for merge, I'll add them.

There's also some discussion needed about field values on targets, which are currently not part of the environment configuration system, it is not clear that we need to support these.