Change `fmt` plugin schema to allow for targetless formatters (#16980)
This PR follow up the `lint` schema change to extend it to `fmt`.
Same idea stands, with the big differences being:
- Formatters Partitions are always of files
- `fmt` has to make "lanes" of disjoint subpartitions to ensure that no parallel execution of formatters formats the same file

Note, there's a subtle contract violation I'll file an issue for where we _might_  split a terraform directory into multiple subpartitions if a subset of files is either larger than the batch size (unlikely) or a file formatter tries to format a strict subset of them. I'll file a ticket where we can discuss. @tdyas has confirmed this isn't actually an issue today (the files will still be formatted the same way), but we should track it nonetheless.

Some future changes:
- partitioner rule maker for the "default" partitioners
- "coersion" rules to handle the plumbed-through params of `LintResult` and `FmtResult`

[ci skip-rust]
[ci skip-build-wheels]