Document how CA certs path works with `docker_environment` and `remote_environment` (#17014)
Closes https://github.com/pantsbuild/pants/issues/16987.

We have three uses for certs path:

1. Connecting to remote execution server
2. The `DownloadFile` intrinsic, which always happens on local host regardless of Docker and remote execution
3. Subprocesses like Pex and Twine

Those first two uses would be difficult to load via environment targets, as they are used when bootstrapping Pants itself. I think I could possibly get it to work: we'd only let you set the values via `local_environment` targets. Every where that calls the helper `determine_bootstrap_environment`, we'd set = the certs path set to `None` at first, then determine the certs path to use, and finally start a new session:

https://github.com/pantsbuild/pants/blob/9769d88d2b914a51ffe7210495fdc5995ca64418/src/python/pants/core/util_rules/environments.py#L302-L307

That would work as long as we don't make network calls to determine the environment target, which we don't. But this all would be difficult to pull off. 

Meanwhile, for subprocesses, it's not hard for us to use an "environment aware" option that env targets can set, so long as we added a dedicated option like `[GLOBAL].ca_certs_path_subprocesses`. But @illicitonion wisely pointed out that should generally not be necessary: we can simply copy the _contents_ of the certs from local host into those environments, which we already do!

So, there's little value to setting certs paths for `docker_environment` and `remote_environment`. While it is valuable to set it for `local_environment`, so that e.g. macOS can have different config than Linux, it is too hard to achieve this given the bootstrapping chicken-and-egg problem. 

Therefore, this PR keeps the status quo, but documents these nuances.

[ci skip-rust]
[ci skip-build-wheels]