[internal] Preprocesses python bootstrap search paths to check for invalid settings (#17041)
Per discussion in #16800, we're OK with using direct filesystem access to support pyenv support (and separately, asdf support) for local runtime environments only. This adds a preprocessing function that supports:

* no-opping for local/undefined environments
* handling the case where the search paths are unaltered from the default, by stripping out unsupported search paths
* raising an exception with fixing instructions if unsupported search paths are requested for remote/docker environments

n.b. `PythonBootstrap` is still incorrectly cached for pyenv users, and that will be addressed in one (hopefully) final PR.