Restore the deprecated `filter` goal indefinitely (#17027)
It's very little code for us to support the `filter` goal, which avoids our users needing to rewrite a bunch of scripts and CI config.

The important thing we wanted is for `./pants help` to render properly, that this is no longer a goal and is now a subsystem. Turns out, `help` was broken before! We special case this goal to handle this properly.

[ci skip-rust]
[ci skip-build-wheels]