Add python 3.10 to acceptable GCF runtimes (Cherry-pick of #17028) (#17053)
Based on https://cloud.google.com/functions/docs/concepts/python-runtime, currently `pants` does not allow you to pick python 3.10. This PR adds that runtime as one of the enums so it doesn't fail anymore.

I have tested this change against another repository that creates a GCF and `pants` was able to successfully generate the .zip file. I also pushed that file to Google Cloud and it worked as expected - tested it by using a `match case` statement. 

[ci skip-rust]

Co-authored-by: Thales Menato <thales.menato@optimizely.com>