Auto-apply reviewers to cherrypicks.
Finds the reviewers of the source PR and applies them to
the cherrypick. Means one less manual step when cherrypicking.

Also checks that a category label was successfully detected.
I've noticed that occasionally one is not, which then causes
CI to fail. This check will allow us to get to the bottom of that.