Introduce `distinct_union_type_per_subclass` for those "special" inner class unions (#17040)
The diff should show the difference in syntax here. The new way not only appeases `mypy` through a one-and-only type definition, but continues to allow for the per-subclass distinct-but-equal semantics.

Note I had to combine the technique with `unions` directly because `@union` stores an attribute pointing to the type itself. When copied this strategy failed. That's OK though because this technique is only useful for unions :smile:  

[ci skip-rust]
[ci skip-build-wheels]