Use a VenvPex when exporting python venvs.
Saves a lot of superfluous zipping and unzipping, so
should speed export up quite a bit.

However note that copying a venv to dist/ takes non-trivial
time for larger venvs, so the symlink trick should be re-introduced
in a followup change.

This change also refactors the export code to unify the resolve
and tool exports, resolving a TODO to that effect.