Optionally export venvs via symlink into the pex_root.
Uses the new plugin subsystem options to make this an
option on the `export` scope, which is neat!

The symlinked venv has no pip binary and so is immutable.
This is dramatically faster than creating a pipified venv,
but any attempt to modify such a venv will corrupt the pants
cache, so caveat emptor.

Also refactors export_test.py to use parametrization.