Always include relevant params in workunit descriptions, and include descriptions in traces by default (#17119)
In order to:
1. begin to render an environment in more places (when one is in use)
2. add more human readable, useful context to errors when they occur

... this change appends all `EngineAwareParams` to the workunit descriptions of tasks, and renders workunit descriptions in engine backtraces by default.

----

The impact on common use is fairly subtle, because this does not impact `Process` (which represents the majority of labeled `debug` level workunits), but you can see it in test output rendering:
```
09:56:14.93 [INFO] Completed: Run Pytest - (environment:docker, src/python/pants/util/strutil_test.py:tests) - succeeded.
```

You can see the impact on backtraces in the changes to the tests. As one example (note: this is with `--no-print-stacktrace`):
```
$ ./pants --no-print-stacktrace dependencies src/python/pants/base/::
10:01:35.84 [ERROR] 1 Exception encountered:

Engine traceback:
  in `dependencies` goal
  in Resolve direct dependencies of target - src/python/pants/base/exception_sink_integration_test.py:exception_sink_integration_test

InvalidSpecPathError: Invalid address /testprojects:pants_plugins_directory from the `dependencies` field from the target src/python/pants/base/exception_sink_integration_test.py:exception_sink_integration_test. Cannot use absolute paths.
```

While this information is lightly redundant with the use cases which were improved in #14468, there are likely to be other errors which currently do not have enough context added. And almost no explicit inclusion of the environment in error messages currently happens, but would need to if context were to be added manually.

[ci skip-build-wheels]