Fix quoting of GitHub Actions expressions. (#17172)
We were previously using ${{ expr }} in f-strings, which
caused them to emit ${ expr } into the YAML, which is incorrect.

Correctly quoting this in an f-string would require 4 braces,
which is unwieldy. Plus, it's common while editing to turn a
regular string into an f-string or vice versa, and it would be
very easy to neglect to change the quoting as needed.

Instead, we introduce a helper function and use it exclusively.

Note that this bug caused us to generate useless, constant keys
for the GHA cache - in fact I noticed something was awry when
I saw CI builds re-building rust code after a successful cache
restore. Good thing our rust bootstrapping script is defensive.
I'd say we were pretty lucky if the only impact was performance;
This could potentially have lead us to actively step on our own
toes in bad ways.