Set PEX_ROOT for export post-processing command. (#17177)
This is both a leak that foils our recommendations to users for maintenance
of Pants caches and results in unneeded extra work when an export finds an
empty ~/.pex but would have had hits in ~/.cache/pants/named_caches/pex_root
if it had been pointed there.

Similar to #17177 but not a straight cherry-pick, as it would not
apply even remotely cleanly.
[ci skip-rust]

[ci skip-build-wheels]

[ci skip-rust]