Use a standard Cargo caching action (#17169)
It will automatically use correct keys and take care of trimming the cache entries.

However, we first had to patch that action to not wipe ~/.cargo/bin, see
Swatinem/rust-cache@master...benjyw:rust-cache:master

I will send that patch, or similar, upstream for consideration, so that in the future we can
use the standard action instead of the forked one.