Refactor `shunit2` into a proper subsystem.
While working on partitioned/batched testing I wanted to add a `skip`
option to all of our existing `test` subsystems - I then noticed that
`shunit2` was an odd-one-out because it didn't have an actual subsystem!

Factoring out the subsystem will enable adding the `skip` option, and
also (hypothetically) allow users to bump to one of the later commits in
the `shunit2` repo.

[ci skip-build-wheels]