Drop requirement for partition key types to have a `description`
Requiring all implementations of partition keys to have a `description`
property has proven to be a rough edge when implementing the
one-partition-per-input strategy. By loosening the restriction we make
it easier for plugin authors to use types like `Address` or `str` as
their partition keys, and also enable the generation/lookup of partition
descriptions within tool-execution rules (instead of forcing it to
happen in partitioning rules). Plugins that _want_ to generate a
description as part of their partitioning rule are still able to do so -
this is demonstrated by a handful of updated plugins in this PR.

This will be convenient when implementing the framework for partitioned
testing.