Simplify temp pex cleanup in Python venv export. (#17317)
Use a truly unique tmpdir for both the exported requirements pex and the pex pex, and delete that dir at the end.

Obviates the need for the `--remove=` flag in the pex invocation, which has caused accidental havoc. 

Generally makes this code easier to reason about, since the name is now a uuid and the deletions happen in one place.