Fix base `specs_integration_test.py`.
For some reason, rearranging the rules in the `test` API has changed the
ouput to no longer match what was expected by these tests (timing info
is now present in the test's nested output, and a dividing `-` has
disappeared). The core information of the lines is the same, and the
output now matches what we assert on in the unit tests for the `test`
goal itself, so I think we're ok just updating the expectations of the
specs test to get things passing again.