Refactor optional typing in generic partitioning utils.
Prior to this commit, the generic partitioning code was typed so:

  * Partitions could have `None` metadata, and
  * If partition metadata was non-`None`, it had to implement a
    `description(self) -> str` property

While working on batched `pytest`, I've found these type constraints to
be a bit too restrictive. Pytest batches will either be:

  * A collection of files all marked as compatible
  * A single file, with no compatibility tag

In the 1st case, we can use the compatibility tag as the partition's
metadata description. In the 2nd case there is no meaningful metadata
description - we could use the single file's address spec, but that
would produce significant noise in our logging. Ideally we could return
`None` in the 2nd case, but the types don't currently allow it because
all the `Partition`s returned by a plugin must have the same type (or be
`None`).

This commit attempts to fix the UX problem by rearranging where we allow
optional values. Metadata is now always required, but the `description`
property is allowed to return `None`. A new "empty" placeholder type is
used by our default rules, so users still won't have to think about it
unless they choose to customize their partitioning logic.